update code to use sensor_device.h
